feat(chat): implement JSON viewer for enhanced data visualization in chat components
- Added a new JsonViewer component to display JSON data in a structured format, allowing users to toggle between code and tree views.
- Integrated JsonViewer into ToolMessage and ToolDetail components for improved data presentation.
- Updated the lazy highlighter theme to 'tomorrow' for better aesthetics.
- Removed redundant TreeNode implementation from ToolMessage, streamlining the component's structure.

These enhancements improve the user experience by providing a clearer and more interactive way to view complex JSON data in the chat interface.
