+++ apps/web/src/components/chat/lazy-highlighter.tsx
@@ -1,6 +1,6 @@
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 // @ts-ignore - correct
-import { materialDark } from "react-syntax-highlighter/dist/esm/styles/prism/index.js";
+import { tomorrow } from "react-syntax-highlighter/dist/esm/styles/prism/index.js";
 
 interface LazyHighlighterProps {
   language: string;
@@ -11,7 +11,7 @@ function LazyHighlighter({ language, content }: LazyHighlighterProps) {
   return (
     <SyntaxHighlighter
       language={language || "text"}
-      style={materialDark}
+      style={tomorrow}
       customStyle={{
         margin: 0,
         padding: "1rem",
